client: explicitly use a 64-bit type for ‘total_len’
This fixes an out-of-range comparison in 32-bit builds:

 client.c:289:19: error: result of comparison of constant 4294967296 with expression of type 'size_t' (aka 'unsigned int') is always false [-Werror,-Wtautological-constant-out-of-range-compare]
        if (total_len >= 1llu << (8 * sizeof(uint32_t)) ||
            ~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ client.c
@@ -263,7 +263,7 @@ main(int argc, char *const *argv)
     };
 
     /* Total packet length, not (yet) including argv[] */
-    size_t total_len = (
+    uint64_t total_len = (
         sizeof(data) +
         cwd_len +
         term_len +
